requirements.txt
streamlit==1.32.0
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.1